Load Inventory Food

+++ edu/ucalgary/ensf409/FoodItem.java
@@ -5,24 +5,24 @@ package edu.ucalgary.ensf409;
 public class FoodItem{
 	private int id;
 	private String name;
-	private int fruitVeggiesContent;
-	private int grainContent;
-	private int proteinContent;
-	private int other;
-	private int calories;
+	private double fruitVeggiesContent;
+	private double grainContent;
+	private double proteinContent;
+	private double other;
+	private double calories;
 	
 	// Getters
 	public int getID() { return this.id; }
 	public String getName() { return this.name; }
-	public int getFruitVeggieContent() { return this.fruitVeggiesContent; }
-	public int getGrainContent() { return this.grainContent; }
-	public int getProteinContent() { return proteinContent; }
-	public int getOther() { return this.other; }
-	public int getCalories() { return this.calories; }
+	public double getFruitVeggieContent() { return this.fruitVeggiesContent; }
+	public double getGrainContent() { return this.grainContent; }
+	public double getProteinContent() { return proteinContent; }
+	public double getOther() { return this.other; }
+	public double getCalories() { return this.calories; }
 	
 	
 	// Constructor
-	public FoodItem(int id, String name, int fvc, int gc, int pc, int other, int cal) {
+	public FoodItem(int id, String name, double fvc, double gc, double pc, double other, double cal) {
 		this.id = id;
 		this.name = name;
 		this.fruitVeggiesContent = fvc;
